Add last tasks

+++ src/task_2/service/CalculatorService.java
@@ -7,15 +7,15 @@ import java.util.Scanner;
 public class CalculatorService {
     private Scanner scanner;
 
-    private static final int X1_1 = -4;
-    private static final int X1_2 = 4;
-    private static final int Y1_1 = 0;
-    private static final int Y1_2 = 5;
+    private static final double X1_1 = -4;
+    private static final double X1_2 = 4;
+    private static final double Y1_1 = 0;
+    private static final double Y1_2 = 5;
 
-    private static final int X2_1 = -6;
-    private static final int X2_2 = 6;
-    private static final int Y2_1 = 0;
-    private static final int Y2_2 = 3;
+    private static final double X2_1 = -6;
+    private static final double X2_2 = 6;
+    private static final double Y2_1 = 0;
+    private static final double Y2_2 = 3;
 
     public CalculatorService() {
         scanner = new Scanner(System.in);
